+++ src/main/java/insane96mcp/insanesurvivaloverhaul/data/generator/ISOItemTagsProvider.java
@@ -1,6 +1,7 @@
 package insane96mcp.insanesurvivaloverhaul.data.generator;
 
 import insane96mcp.insanesurvivaloverhaul.InsaneSO;
+import insane96mcp.insanesurvivaloverhaul.module.farming.crops.Crops;
 import insane96mcp.insanesurvivaloverhaul.module.hungerhealth.fooddrinks.FoodDrinks;
 import insane96mcp.insanesurvivaloverhaul.module.items.copper.CopperEquipment;
 import insane96mcp.insanesurvivaloverhaul.module.misc.tweaks.Tweaks;
@@ -90,6 +91,9 @@ public class ISOItemTagsProvider extends ItemTagsProvider {
         tag(TerrainSlowdown.SNOW_SLOWDOWN_IGNORE)
                 .add(Items.LEATHER_BOOTS)
                 .addOptional(ResourceLocation.parse("tconstruct:travelers_boots"));
+
+        tag(Crops.CHICKEN_FOOD_ITEMS)
+                .add(Items.MELON_SEEDS, Items.PUMPKIN_SEEDS, Items.TORCHFLOWER_SEEDS);
     }
 
     public static TagKey<Item> create(String tagName) {