+++ src/main/java/insane96mcp/insanesurvivaloverhaul/mixin/accessor/VillagerAccessor.java
@@ -1,0 +1,18 @@
+package insane96mcp.insanesurvivaloverhaul.mixin.accessor;
+
+import net.minecraft.world.entity.npc.Villager;
+import net.minecraft.world.item.Item;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.Mutable;
+import org.spongepowered.asm.mixin.gen.Accessor;
+
+import java.util.Set;
+
+@Mixin(Villager.class)
+public interface VillagerAccessor {
+	@Accessor("WANTED_ITEMS")
+	@Mutable
+	static void setWantedItems(Set<Item> requestedItems) {
+		throw new AssertionError();
+	}
+}